% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mv_Poisproc_reg.R
\name{mv_Poisproc_reg}
\alias{mv_Poisproc_reg}
\title{Implementation of the Poisson fSuSiE
# using the Poisson multiscale approach use in multiseq by Shim et al 2024}
\usage{
mv_Poisproc_reg(
  Y,
  Z,
  X,
  L = 3,
  L_start = 3,
  reflect = FALSE,
  verbose = TRUE,
  n_gh = 10,
  init_b_pm,
  tol = 0.001,
  tol_vga_pois = 1e-05,
  maxit = 10,
  control_mixsqp = list(verbose = FALSE, eps = 1e-06, numiter.em = 4),
  thresh_lowcount = 0.01,
  prior_mv = "mixture_normal_per_scale",
  gridmult = sqrt(2),
  nullweight.mrash = 10,
  init_pi0_w.mrash = 10,
  cov_lev = 0.95,
  min_purity = 0.5,
  greedy = TRUE,
  backfit = TRUE,
  tol.mrash = 0.001,
  verbose.mrash = TRUE,
  maxit.mrash = 10,
  cal_obj.mrash = FALSE,
  maxit.fsusie = 50,
  cal_obj.fsusie = FALSE,
  max_SNP_EM = 100,
  max_step_EM = 1,
  cor_small = TRUE,
  max.iter = 3
)
}
\description{
Implementation of the Poisson fSuSiE
# using the Poisson multiscale approach use in multiseq by Shim et al 2024
}
