% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VA_Neg_Binomial.R
\name{nb_mean_lower_bound}
\alias{nb_mean_lower_bound}
\title{Solve Negative Binomial mean problem by Jaakkola Jordan bound}
\usage{
nb_mean_lower_bound(
  x,
  r = 1000,
  ebnm_params = NULL,
  tol = 1e-05,
  maxiter = 1000,
  est_r = FALSE,
  update_r_every = 1,
  m_init = NULL,
  v_init = NULL
)
}
\arguments{
\item{x}{data vector}

\item{r}{NB binomial paramter r}

\item{ebnm_params}{a list of ebnm parameters}

\item{tol}{tolerance for stopping the updates}

\item{maxiter}{max number of iterations}

\item{est_r}{whether estimate r. For Poisson data, set it to False.}

\item{update_r_every}{if estimate r, how often to update it}
}
\value{
a list of
 \item{posteriorMean:}{posterior mean}
 \item{v:}{posterior variance}
 \item{obj:}{objective function values}
 \item{ebnm_res:}{fitted object from `ebnm`}
 @example
 n = 10000
 mu = rnorm(n)
 x = rpois(n,exp(mu))
 nb_mean_lower_bound(x)
}
\description{
Solve Negative Binomial mean problem by Jaakkola Jordan bound
}
\details{
The problem is
\deqn{x_i\sim Poisson(\log(1+\exp(\mu_i))),}
\deqn{\mu_i\sim g(\cdot).}
}
