% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Poisson_Gaussina_posterior.R
\name{pois_mean_GG1_obj_gradient}
\alias{pois_mean_GG1_obj_gradient}
\title{calculate gradient vector}
\usage{
pois_mean_GG1_obj_gradient(theta, x, s = NULL, beta, sigma2)
}
\description{
calculate gradient vector
}
