% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Poisson_Gaussina_posterior.R,
%   R/nugget_split_VA.R
\name{pois_mean_GG_opt_obj_gradient}
\alias{pois_mean_GG_opt_obj_gradient}
\title{calculate gradient vector}
\usage{
pois_mean_GG_opt_obj_gradient(theta, x, beta, sigma2, n)

pois_mean_GG_opt_obj_gradient(theta, x, beta, sigma2, n)
}
\description{
calculate gradient vector

calculate gradient vector
}
