% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Poisson_Gaussina_posterior.R
\name{pois_mean_GP_opt_obj_gradient}
\alias{pois_mean_GP_opt_obj_gradient}
\title{calculate gradient vector}
\usage{
pois_mean_GP_opt_obj_gradient(theta, x, s, beta, sigma2, n)
}
\description{
calculate gradient vector
}
